feat: Rewrite computer guess logic to use minimax
This iteration removes the previous guess logic which was just a random
picker to the intended guess logic. The new guess logic use two methods
to process each guess, the remove_opts method removes all options that
return the same hints results and the minimax method will use this to
come up with the next best guess.